+++ Magic/CHANGELOG.md
@@ -8,6 +8,7 @@
    - Allow using disabled icons and skull-based disabled icons.
    - Disabled icons will show in skill selector for unavailable skills
    - Can't take an unavailable skill item out of the skill selector
+   - Hotbar skill timers will reflect mana costs
 
 ## 6.6